+++ projects/habitat_uncertainty/run.py
@@ -9,13 +9,13 @@ import torch
 from habitat import get_config
 from habitat.config import read_write
 from habitat.config.default_structured_configs import register_hydra_plugin
-from habitat_uncertainity.config import HabitatConfigPlugin
+from habitat_uncertainty.config import HabitatConfigPlugin
 
 from habitat_baselines.run import execute_exp
-from habitat_uncertainity.task.sensors import YOLOObjectSensor, YOLOStartReceptacleSensor, YOLOGoalReceptacleSensor
-from habitat_uncertainity.utils.YOLO_pred import YOLOPerception as YOLO_pred
-from habitat_uncertainity.models.yoloPointNavResNetPolicy import yoloPointNavResNetPolicy, yoloResNetEncoder
-from habitat_uncertainity.trainers.ppo_trainer_yolo import PPOyoloTrainer
+from habitat_uncertainty.task.sensors import YOLOObjectSensor, YOLOStartReceptacleSensor, YOLOGoalReceptacleSensor
+from habitat_uncertainty.utils.YOLO_pred import YOLOPerception as YOLO_pred
+from habitat_uncertainty.models.yoloPointNavResNetPolicy import yoloPointNavResNetPolicy, yoloResNetEncoder
+from habitat_uncertainty.trainers.ppo_trainer_yolo import PPOyoloTrainer
 
 def register_plugins():
     register_hydra_plugin(HabitatConfigPlugin)
